+++ src/Facades/openedx.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace ngunyimacharia\openedx\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+class openedx extends Facade
+{
+    /**
+     * Get the registered name of the component.
+     *
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return 'openedx';
+    }
+}