+++ src/Listeners/UserEventSubscriber.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace ngunyimacharia\openedx\Listeners;
+
+class UserEventSubscriber
+{
+    /**
+     * Handle user login events.
+     */
+    public function handleUserLogin($event) {
+        dd($event);
+    }
+
+    /**
+     * Handle user logout events.
+     */
+    public function handleUserLogout($event) {}
+
+    /**
+     * Register the listeners for the subscriber.
+     *
+     * @param  \Illuminate\Events\Dispatcher  $events
+     */
+    public function subscribe($events)
+    {
+        $events->listen(
+            'Illuminate\Auth\Events\Login',
+            'ngunyimacharia\openedx\Listeners\UserEventSubscriber@handleUserLogin'
+        );
+
+        $events->listen(
+            'Illuminate\Auth\Events\Logout',
+            'ngunyimacharia\openedx\Listeners\UserEventSubscriber@handleUserLogout'
+        );
+    }
+}