+++ src/laravel-openedx.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace ngunyimacharia\laravel-openedx;
-
-class laravel-openedx
-{
-    // Build wonderful things
-}