+++ src/openedx.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ngunyimacharia\openedx;
+
+class openedx
+{
+    // Build wonderful things
+}