+++ changelog.md
@@ -1,8 +1,14 @@
 # Changelog
+All notable changes to this project will be documented in this file.
 
-All notable changes to `laravel-openedx` will be documented in this file.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Version 1.0
+
+## 0.0.0 - 2019-04-29
 
 ### Added
-- Everything
+- Created Listeners, Models, Observers, Service Providers and openedx.php
+
+
+[0.0.0]: https://github.com/ngunyimacharia/innovativkonzept/releases/tag/v0.0.0